Cleanup logs_test.go
Removed a lot of extraneous tests because they were effectively testing
the same situations.

On windows, closing the server too soon causes an error on the client.
FOr now, we've decided to paper over this by increasing the sleep but
finding an isolated testcase and reporting this to the mailing list is
the next step.