FakeAppInstancesRepo should be easier to use
The behavior before was that each time you called GetInstances, it would slice off
one of the responses and errors from its slick of mock responses. Eventually there
would be nothing left and it would just return you the zero-value for the
response and error, which made it fairly difficult to use this in practice
because we poll this repo over and over again in practice.