Use '?' instead of '-1' when running instances is unknown
When the health manager is unable to determine the desired or actual
running instances of an application, the cloud controller will present a
-1 value for running instances to the caller.  This is a relatively
recent behavior change.

To more accurately represent the unknown state, use a '?' in the cli
output for running instances when the value from the cloud controller is
negative.